requirements.txt
tensorflow-gpu==1.12.0
tensor2tensor==1.14.0 
numpy==1.17.4
librosa==0.7.1 
speechpy
tqdm
sox
